Updated README.md Added some extra text for empty lists in course_list More generic title
