+++ src/course_management/templates/base.html
@@ -4,7 +4,7 @@
   <link type="text/css" rel="stylesheet" href="{% static "vendor/bootstrap/dist/css/bootstrap.min.css" %}">
   <link type="text/css" rel="stylesheet" href="{% static "css/layout.css" %}">
 	<link href='https://fonts.googleapis.com/css?family=Exo+2' rel='stylesheet' type='text/css'>
-  <title>{% if title %}{{ title }} | iFSR Course Management{% else %}iFSR Course Management{% endif %}</title>
+  <title>{% if title %}{{ title }} | {% endif %}Course Management</title>
 </head>
 <body class="container">
   {% block header %}
